@@ -8,26 +8,25 @@
   "license": "MIT",
   "type": "project",
   "require": {
-    "php": "^7.4",
-    "laravel/framework": "^7.0",
-    "laravel/tinker": "~2.0",
+    "php": "^8.1",
+    "laravel/framework": "^8.0",
+    "laravel/tinker": "^2.0",
     "symfony/translation-contracts": "^2.0.1",
     "stripe/stripe-php": "^7.14"
   },
   "require-dev": {
-    "fzaninotto/faker": "~1.4",
-    "mockery/mockery": "~1.0",
-    "phpunit/phpunit": "~8.0",
-    "symfony/css-selector": "~5.0",
-    "symfony/dom-crawler": "~5.0",
-    "filp/whoops": "~2.0"
+    "fakerphp/faker": "^1.17",
+    "mockery/mockery": "^1.0",
+    "phpunit/phpunit": "^9.3",
+    "symfony/css-selector": "^5.0",
+    "symfony/dom-crawler": "^5.0",
+    "filp/whoops": "^2.0"
   },
   "autoload": {
-    "classmap": [
-      "database"
-    ],
     "psr-4": {
-      "App\\": "app/"
+      "App\\": "app/",
+      "Database\\Factories\\": "database/factories/",
+      "Database\\Seeders\\": "database/seeders/"
     }
   },
   "autoload-dev": {
