@@ -4,7 +4,9 @@ namespace Tests\Unit\Billing;
 
 use App\Billing\PaymentFailedException;
 use App\Billing\PaymentGateway;
-
+/**
+ * @group Unit
+ */
 trait PaymentGatewayContractTest
 {
     abstract protected function getPaymentGateway(): PaymentGateway;
